Renamed viewer filenames.
