+++ catkin.cmake
@@ -69,12 +69,12 @@ add_library(${PROJECT_NAME}_rviz SHARED
 
 # OpenRAVE viewer classes.
 qt4_wrap_cpp(VIEWER_MOC
-    include/${PROJECT_NAME}/or_rviz.h
+    include/${PROJECT_NAME}/RVizViewer.h
 )
 
 add_library(${PROJECT_NAME} SHARED
-    src/or_interactivemarker.cpp
-    src/or_rviz.cpp
+    src/InteractiveMarkerViewer.cpp
+    src/RVizViewer.cpp
     ${VIEWER_MOC}
 )
 target_link_libraries(${PROJECT_NAME}
